net: warn if only loopback nameservers found
On many systems today the nameserver functionality is handled by an
application such as systemd-resolved or dnsmasq and the nameserver
address is a loopback address (like 127.0.0.53) where that application
listens for incoming DNS queries.

This will not work for us in this network namespace environment, since
we have our own namespaced loopback device.