Add --use-public-dns option
This commits adds the --use-public-dns command line option.

In order to have our own nameservers for our network namespace we will
bind mount our own resolv.conf over the system-wide one at
/etc/resolv.conf.

But we do not want to affect others outside this namespace. So we
enter a mount namespace (CLONE_NEWNS) before we do the bind mount. This,
however, is not enough. Before the bind mount we need to remount the
root partition with the MS_SLAVE and MS_REC flags to make sure our
changes are not propagated to the outside mount namespace. As some
systems have that as the default behavior.