Rework network interface names and IP allocation
In order to be able to run multiple instances of nsntarce we need to
avoid clashing in names of interfaces and in the allocation of IP
addresses.

We solve this by using the PID of the nsntrace namespace process
in the network interface names. And by checking already bound IP
address and avoid them. We use randomization when allocating
an IP address to check, to avoid race conditions when launcing many
nsntrace instances at the same time.

There are still some risk for collision when instances launch within
the same millisecond.